fix(ai31): restore source-size lint compliance
