+++ crates/atm-core/src/send/mod.rs
@@ -43,6 +43,7 @@ pub mod input;
 #[doc(hidden)]
 pub(crate) mod nudge_template;
 mod persistence;
+mod recipient;
 pub(crate) mod summary;
 
 pub(crate) use delivery_persistence::{
@@ -55,6 +56,7 @@ pub use nudge_template::{
 };
 #[cfg(test)]
 pub(crate) use persistence::persist_message;
+pub(crate) use recipient::{ResolvedRecipient, resolve_recipient, validate_non_self_recipient};
 
 pub(super) const POST_SEND_HOOK_TIMEOUT: Duration = Duration::from_secs(5);
 
@@ -1051,146 +1053,6 @@ fn emit_send_command_event(
     }) {
         warn!(%error, command = "send", action = "send", "failed to emit send command event");
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct ResolvedRecipient {
-    pub(crate) agent: AgentName,
-    pub(crate) team: TeamName,
-}
-
-pub(crate) fn validate_non_self_recipient(
-    sender: &AgentName,
-    sender_team: &TeamName,
-    recipient: &ResolvedRecipient,
-    target: &AgentAddress,
-    provenance: ValidatedWriteProvenance,
-) -> Result<(), AtmError> {
-    let same_identity = sender
-        .as_str()
-        .eq_ignore_ascii_case(recipient.agent.as_str())
-        && sender_team
-            .as_str()
-            .eq_ignore_ascii_case(recipient.team.as_str());
-    if same_identity && target.host().is_none() && !provenance.is_authenticated_peer() {
-        return Err(AtmError::self_addressed_send_invalid(format!(
-            "self-addressed messages are invalid ATM input: '{sender}@{sender_team}' may not send to itself"
-        )));
-    }
-    Ok(())
-}
-
-#[cfg(test)]
-mod self_address_tests {
-    use super::{ResolvedRecipient, validate_non_self_recipient};
-    use crate::address::AgentAddress;
-    use crate::error_codes::AtmErrorCode;
-    use crate::provenance::{WriteIngress, WriteProvenance, validate_write_provenance};
-    use crate::types::{AgentName, TeamName};
-
-    #[test]
-    fn validate_non_self_recipient_rejects_case_variant_self_target() {
-        let provenance = validate_write_provenance(
-            WriteIngress::Canonical,
-            WriteProvenance {
-                target_host: None,
-                authenticated_source_host: None,
-                origin_message_id: false,
-                origin_timestamp: false,
-            },
-        )
-        .expect("local provenance");
-        let error = validate_non_self_recipient(
-            &AgentName::from_validated("Sender-A"),
-            &TeamName::from_validated("Test-Team"),
-            &ResolvedRecipient {
-                agent: AgentName::from_validated("sender-a"),
-                team: TeamName::from_validated("test-team"),
-            },
-            &"sender-a@test-team"
-                .parse::<AgentAddress>()
-                .expect("target"),
-            provenance,
-        )
-        .expect_err("case-variant self target must be rejected");
-
-        assert_eq!(error.code(), AtmErrorCode::SelfAddressedSendInvalid);
-    }
-
-    #[test]
-    fn validate_non_self_recipient_allows_host_qualified_self_target() {
-        let target = "sender-a@test-team.127.0.0.1"
-            .parse::<AgentAddress>()
-            .expect("host-qualified target");
-        let provenance = validate_write_provenance(
-            WriteIngress::Canonical,
-            WriteProvenance {
-                target_host: target.host(),
-                authenticated_source_host: None,
-                origin_message_id: false,
-                origin_timestamp: false,
-            },
-        )
-        .expect("host-qualified origin provenance");
-        validate_non_self_recipient(
-            &AgentName::from_validated("sender-a"),
-            &TeamName::from_validated("test-team"),
-            &ResolvedRecipient {
-                agent: AgentName::from_validated("sender-a"),
-                team: TeamName::from_validated("test-team"),
-            },
-            &target,
-            provenance,
-        )
-        .expect("host-qualified self target must use the ordinary peer route");
-    }
-
-    #[test]
-    fn validate_non_self_recipient_allows_authenticated_peer_after_target_normalization() {
-        let target = "sender-a@test-team"
-            .parse::<AgentAddress>()
-            .expect("normalized target");
-        let peer_host = "peer.example.test".parse().expect("peer host");
-        let provenance = validate_write_provenance(
-            WriteIngress::Canonical,
-            WriteProvenance {
-                target_host: target.host(),
-                authenticated_source_host: Some(&peer_host),
-                origin_message_id: true,
-                origin_timestamp: true,
-            },
-        )
-        .expect("authenticated peer provenance");
-        validate_non_self_recipient(
-            &AgentName::from_validated("sender-a"),
-            &TeamName::from_validated("test-team"),
-            &ResolvedRecipient {
-                agent: AgentName::from_validated("sender-a"),
-                team: TeamName::from_validated("test-team"),
-            },
-            &target,
-            provenance,
-        )
-        .expect("authenticated peer receipt must not become a local self-send");
-    }
-}
-
-fn resolve_recipient(
-    target_address: &AgentAddress,
-    caller_team: &TeamName,
-    config: Option<&config::AtmConfig>,
-) -> Result<ResolvedRecipient, AtmError> {
-    // `AgentAddress` has already validated the explicit team segment. Never
-    // parse it again and silently substitute the caller team on failure.
-    let team = target_address
-        .team()
-        .cloned()
-        .unwrap_or_else(|| caller_team.clone());
-
-    Ok(ResolvedRecipient {
-        agent: config::aliases::resolve_agent_name(target_address.agent(), config)?,
-        team,
-    })
 }
 
 fn resolve_message_body(
